# backend/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products/{productId}/reviews": {
      "post": {
        "description": "買家對購買的商品進行評論",
        "operationId": "ReviewController_createReview",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "商品ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "評論創建成功"
          },
          "400": {
            "description": "請求參數錯誤或尚未購買該商品"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "商品不存在"
          },
          "409": {
            "description": "已評論過該商品"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "創建商品評論",
        "tags": [
          "商品評論"
        ]
      },
      "get": {
        "description": "獲取指定商品的所有評論",
        "operationId": "ReviewController_getProductReviews",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "商品ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每頁數量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "頁碼",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取商品評論成功"
          },
          "404": {
            "description": "商品不存在"
          }
        },
        "summary": "獲取商品評論",
        "tags": [
          "商品評論"
        ]
      }
    },
    "/api/v1/reviews/{reviewId}": {
      "patch": {
        "description": "更新用戶自己的評論內容",
        "operationId": "ReviewController_updateReview",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "description": "評論ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "評論更新成功"
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足或不是評論作者"
          },
          "404": {
            "description": "評論不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新評論",
        "tags": [
          "商品評論"
        ]
      },
      "delete": {
        "description": "刪除用戶自己的評論或管理員刪除任意評論",
        "operationId": "ReviewController_deleteReview",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "description": "評論ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "評論刪除成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足或不是評論作者"
          },
          "404": {
            "description": "評論不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "刪除評論",
        "tags": [
          "商品評論"
        ]
      }
    },
    "/api/v1/products/{productId}/images": {
      "post": {
        "description": "為指定商品上傳圖片（最多8張）",
        "operationId": "ImageController_uploadImages",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "商品ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "圖片上傳成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Images uploaded successfully."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "filename": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "沒有上傳檔案或檔案格式錯誤"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "商品不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "上傳商品圖片",
        "tags": [
          "圖片管理"
        ]
      }
    },
    "/api/v1/images/{imageId}": {
      "delete": {
        "description": "刪除指定的商品圖片",
        "operationId": "ImageController_deleteImage",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "description": "圖片ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "圖片刪除成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足或不是圖片擁有者"
          },
          "404": {
            "description": "圖片不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "刪除圖片",
        "tags": [
          "圖片管理"
        ]
      }
    },
    "/api/v1/imagesFromName/{imgName}": {
      "get": {
        "description": "根據檔案名稱獲取圖片資訊",
        "operationId": "ImageController_getImageByFilename",
        "parameters": [
          {
            "name": "imgName",
            "required": true,
            "in": "path",
            "description": "圖片檔案名稱",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "圖片獲取成功"
          },
          "404": {
            "description": "圖片不存在"
          }
        },
        "summary": "根據檔名獲取圖片",
        "tags": [
          "圖片管理"
        ]
      }
    },
    "/api/v1/imagesFromID/{imageId}": {
      "get": {
        "description": "根據圖片ID獲取圖片資訊",
        "operationId": "ImageController_getImageById",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "description": "圖片ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "圖片獲取成功"
          },
          "404": {
            "description": "圖片不存在"
          }
        },
        "summary": "根據ID獲取圖片",
        "tags": [
          "圖片管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zipperoo 電商平台 API",
    "description": "Zipperoo 電商平台後端 API 文檔",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateReviewDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}